% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rasterfaster.R
\name{resampleBy}
\alias{resampleBy}
\alias{resampleTo}
\title{Resample a numeric RasterLayer}
\usage{
resampleBy(x, factor, method = c("bilinear", "ngb"))

resampleTo(x, nrow = 180, ncol = 360, method = c("bilinear", "ngb"))
}
\arguments{
\item{x}{RasterLayer object to be resampled. Currently it MUST be backed by a
.grd file and must have \code{numeric} data.}

\item{factor}{Factor to resize by (for example, \code{0.5} for 50\%,
\code{3.2} for 320\%).}

\item{method}{\code{"bilinear"} for bilinear interpolation, or \code{"ngb"}
for nearest-neighbor.}

\item{nrow,ncol}{Number of rows and columns in the output layer.}
}
\value{
Resampled raster.
}
\description{
Resample a numeric RasterLayer
}
\examples{
library(raster)
src <- raster(system.file("external/test.grd", package="raster"))
system.time(result <- resampleBy(src, 0.63))
plot(result)
}

