% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findMode}
\alias{findMode}
\title{Find the mode for a vector}
\usage{
findMode(x)
}
\arguments{
\item{x}{An integer, real, character, or logical vector}
}
\description{
Calculates the mode for integer, real, character, and logical vectors. In
the event of a tie, a winner is chosen at random (using C's \code{rand()},
so not affected by \code{\link[base:set.seed]{set.seed()}}, sorry!).
}
\seealso{
Intended to be a faster, less flexible replacement for
  \code{\link[raster:modal]{raster::modal}}.
}

